Split cli, persist running scenarios
* better code organisation
* persist running scenarios to disk for later restore
