+++ src/warnet/warnet.py
@@ -17,7 +17,7 @@ from services.node_exporter import NodeExporter
 from services.grafana import Grafana
 from services.tor import Tor
 from services.fork_observer import ForkObserver
-from services.fluentd import FLUENT_CONF, Fluentd, FLUENT_IP
+# from services.fluentd import FLUENT_CONF, Fluentd, FLUENT_IP
 from services.dns_seed import DnsSeed, ZONE_FILE_NAME, DNS_SEED_NAME
 from warnet.tank import Tank
 from warnet.utils import parse_bitcoin_conf, gen_config_dir, bubble_exception_str
@@ -44,7 +44,7 @@ class Warnet:
             f"copying config {TEMPLATES / FO_CONF_NAME} to {self.fork_observer_config}"
         )
         shutil.copy(TEMPLATES / FO_CONF_NAME, self.fork_observer_config)
-        shutil.copy(TEMPLATES / FLUENT_CONF, self.config_dir)
+        # shutil.copy(TEMPLATES / FLUENT_CONF, self.config_dir)
 
     def __str__(self) -> str:
         tanks_str = ",\n".join([str(tank) for tank in self.tanks])
@@ -177,7 +177,7 @@ class Warnet:
         result = seeder.exec_run(
             f"sh -c 'echo \"{content_str}\" > /etc/bind/dns-seed.zone'"
         )
-        logging.debug(f"result of updating {ZONE_FILE_NAME}: {result}")
+        logger.debug(f"result of updating {ZONE_FILE_NAME}: {result}")
 
         # Reload that single zone only
         seeder.exec_run("rndc reload dns-seed")
